Add files via upload
Regularization is a technique used in machine learning to prevent overfitting by adding a penalty to the model's complexity. It is particularly useful when dealing with models that are prone to overfitting, such as linear models with many features or deep neural networks. By penalizing large coefficients, regularization helps to create simpler models that generalize better to unseen data.
Types of Regularization:
1. L1 Regularization (Lasso Regression)
2. L2 Regularization (Ridge Regression)

How Regularization Works:
Regularization works by adding a regularization term to the loss function (cost function) that the model is trying to minimize. This term penalizes large coefficients, discouraging the model from becoming too complex.